fix(04): keep localized FAQ paths out of client bundle
